# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  version: 1.0.0
  description: |
    Minimal REST API to create orders and read their status.
    Orchestration is handled by Temporal; statuses are persisted in Postgres.
servers:
  - url: http://localhost:3000
paths:
  /orders:
    post:
      summary: Create order and start processing workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              amount: 1000
              customer_id: cust-1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "400":
          description: Bad request
        "500":
          description: Internal error
  /orders/{id}:
    get:
      summary: Get order status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "404":
          description: Not found
        "500":
          description: Internal error
components:
  schemas:
    Status:
      type: string
      enum:
        - created
        - payment_pending
        - payment_confirmed
        - shipping
        - completed
        - failed
        - cancelled
    CreateOrderRequest:
      type: object
      properties:
        amount:
          type: integer
          format: int32
        customer_id:
          type: string
      required: [amount, customer_id]
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/Status'
        reason:
          type: string
          nullable: true
        amount:
          type: integer
          format: int32
        customer_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time